refactor(agent): neutralize VM recovery footprint contract
