+++ packages/agent/test/vm-recovery-footprint-bridge-adversarial.test.ts
@@ -8,11 +8,11 @@ import { describe, expect, it, vi } from 'vitest';
 import {
   enrichVmRecoveryFootprints,
   planVmRecoveryMicrobatch,
-  type VmRecoveryChainFootprint,
   type VmRecoveryFootprintBridge,
   type VmRecoveryMicrobatchLimits,
   type VmRecoveryUpdateContext,
 } from '../src/vm-recovery-microbatch.js';
+import type { VmRecoveryChainFootprint } from '../src/vm-recovery-types.js';
 
 interface TestChainReader {
   isContextGraphActiveOnChain(contextGraphId: bigint): Promise<boolean>;
